Simplify type conversion

+++ pkg/clusteragent/autoscaling/workload/external/recommender_client.go
@@ -139,7 +139,7 @@ func (r *recommenderClient) buildWorkloadRecommendationRequest(clusterName strin
 			targetType = "custom-query"
 			if u := objective.CustomQueryObjective.Value.AbsoluteValue; u != nil {
 				// custom query uses absolute numeric value; convert to float
-				targetValue = float64(int32(u.AsDec().UnscaledBig().Int64()))
+				targetValue = float64(u.AsDec().UnscaledBig().Int64())
 			}
 		}
